Split export code out of votes
